app/requirements.txt
PyYAML~=5.4
APScheduler~=3.7.0
eyeD3~=0.9.4
